SCL-1024 Extend type conformance checking to consider ScProjectionType on the LHS. On both the LHS and RHS, recurse into the StdType wrapped up in:
a@ScProjectionType.element, 
a = Some(b@ScSyntheticClass)
b.t.

Did the same in ScMethodCallImpl to expose a the type wrapped in an ScProjectionType as the return type of the method call.

This implementation should be reviewed -- there may be a better way to implement.

Added a type inference test cases for implicit method calls. Added a new base class for testing type conformance, and a few test cases related to this issue. Many more test cases are required to get good coverage.

Changed the type of ScProjectionType.element from the inferred type Any to Option[PsiNamedElement].

10/700 tests are failing or giving warnings, but I think these were already like this.

git-svn-id: http://svn.jetbrains.org/idea/Trunk/scala@25126 59f64563-22e8-0310-8d2c-98a15e1539f2