Add a test to use inference of the generic type of an implicit conversion.
Convert arrays to immutable.List to avoid RuntimeException when calling hashCode() of ScCompoundType. This is step #1 of the proposed solutions from my email below:



Basically, collections from scala.mutable and arrays (scala.runtime.BoxedArray), throw a runtime error if you call hashcode. There is a good explanation and criticism of the behaviour here: http://scalide.blogspot.com/2009/05/hashcode-equals-in-scala-28-collections.html

This means it is dangerous to have a signature for a case class like:

case class Foo(seq: Seq[Int])

I cannot see a straightforward way to use the type system to say "seq: ImmutableSeq" -- mutable/immutable is only in the package structure, not in the type system.

What should we do about this, to avoid runtime exceptions?
1. Wait for the RuntimeExceptions, and then call Seq.toList where required?
2. Change the signature of case classes intended as hash keys to take immutable.List[A] rather than Seq[A]?
3. Something else?

-jason




git-svn-id: http://svn.jetbrains.org/idea/Trunk/scala@25754 59f64563-22e8-0310-8d2c-98a15e1539f2