SCL-2872 Centralize and improve checks for whether a type is FunctionN / PartialFunction / TupleN.
Dealias type before checking if it is TupleN/Function/PartialFunctionN.

Types hidden behind higher kinded type aliases are now correctly substituted. The previous approach in ScParameterizedType#getFunctionType, of assuming the typeArgs could be directly applied to FunctionN, was insufficent for:

      type F[A] = (A => Any); { case x => x }: F[Int].

Instead, we use a substitutor to fill in the correct type args.

ScType.expandAliases needs a bit more polish -- it satisfies the test cases I've added, but I don't think it is comprehensive.