SCL-2697 First cut at apply/unapply find usages.
Only works in the simplest cases, if the method is in an object, and that object appears at the call-site of apply/unapply.

object A {
  def apply(a: Any) = ()
  def unapply(a: Any) = false
}

A(0)

0 match {
  case A() =>
}

The next step seems like a brute-force search. Maybe in the meantime we should issue a warning about find-usages or rename of these methods.