Primary constructors now get the full 'effective parameter clauses' treatment.
An empty param clause is prepended to ensure at least one non-implicit clause. View and context bounds are added as a synthetic implicit parameter clause.

Secondary constructors inherit the synthetic param clause from the primary constructor.

Add tests to demonstrate this working, for primary and secondary constructors; and for functions.