Minimal support for macro defs.
These are available in Scala trunk, behind -Xmacros. The plugin allows these irrespective of this setting.

`macro` is not a keyword, so the lexer passes it as an identifier. FunDef now can have two identifiers if the first one is "macro".

  def macro foo() = Constant(Literal(0)
  def macro macro(a: Any*): Boolean = Constant(Literal(0)

Future work: inside the body of a macro def, resolve according to the desugaring described:

  https://github.com/scala/scala/blob/master/src/compiler/scala/tools/nsc/typechecker/Macros.scala#L29