Add a synthetic pre-amble to the body of a macro def.
The body of a macro def sees the parameters and type parameters as values of type _context.Tree, and has two extra parameters in provded: _context: Context, and _this: Tree. This is simulated here by ignoring the parameters, and adding a pre-amble to the function body. This approach seemed more straight forward than switching in a synthetic parameter list.

The expected type of a macro def is `_context.Tree`, but for now we just use Any.