HOCON include resolution and accompanying refactors
This commit implements resolution of HOCON `include` statements. Quoted strings inside includes now contain references to files found in sources and libraries. Also, there is an inspection for unresolved includes.

Details and assumptions:
* Only `classpath(...)` and generic includes are supported. Thus, `file(...)` and `url(...)` includes are not supported.
* Including file is assumed to have been loaded from classpath. Therefore, each generic include is interpreted as a relative path to other classpath resources (unless it's explicitly absolute - starts with a slash)
* `classpath(...)` includes are interpreted as absolute paths to classpath resources.
* If including file is in the sources of a module, included files are searched in entire classpath of this module (sources, dependencies and libraries).
* If including file is in a library, included files are searched in merged classpath of all modules that directly depend on that library.
* Test sources and test dependencies are searched only if including file is also a test source or lies inside a test dependency.
* If included file name doesn't have one of `.conf`, `.json` or `.properties` extensions, the file name will be appended with all of these extensions and each will be searched for. Therefore, there is no way to include file with any other extension. This is also reflected by an appropriate completion filter.

The scope where included files are searched for is potentially the most controversial issue. We have no way of knowing with what classpath the HOCON file will be loaded in runtime, so we don't know exactly where to search. The actual scope may be much wider.

Refactors and general improvements that come along with include resolution implementation:
* extended API of some HOCON PSI elements
* introduced a dedicated visitor class for HOCON PSI elements
* various string constants moved to common object
* cosmetic renames

Implementation of include resolution is based on code found in `FilePathReferenceProvider`, `FileReferenceSet`, `FileReference` and implementations of `FileReferenceHelper` - primarily the `PsiFileReferenceHelper`. Code of these components is reused as much as possible.