Fix processing weight for ScalaProjectData
Since SbtProjectDataService (which processes ScalaProjectData) needs
information about modules, ScalaProjectData should have higher processing weight
than ProjectKeys.MODULE.

It solves the following problem: imagine a project that is being imported for the
first time. We have all its modules as DataNodes, but there are no modules
in IDEA project yet. Suppose the project has a shared source module. Then we
need to set SBT incrementality type for the whole project, but we won't. It is
becuase ScalaProjectData will be processed before modules DataNodes and since
there are no modules in IDEA yet it will skip the step where we set up
incrementality type based on existence of shared source module in a project.