Refactor simple importing test using ProjectStructureDsl
Also Implement different strategies of property comparison.

ExactMatch has behaviour similar to assertSomething functions from
ExternalSystemImportingTestCase: it asserts that things are matching only when
actual value is exactly the same as expected and vice versa.

InexactMatch looses this condition on sequences, allowing actual sequence to
contain more items than are in the expected one. This is very handy when we
have a project where we interested in some particular library or content root
to be imported and have no strong feelings for others.