fix choose var or val when work with uninitalize final variable that initalize not with constructor param.
When final variable is not initalize in declaration we can initalize it only in constructor. If there is only one usage and variable handle as param to constructor - initalize as constructor params, but when variable is not initalize with constructor param we got val instead var (ok. finals should be val, but if we can't choose primary constructor  - we make it var)