Introduce BaseScalaApplicationConfigurationProducer
The goal of this trivial refactoring is to be able to provide in
plugins new types of configurations and configuration types extending
the existing ones that are available via a context menu as separate
entries always when the current Scala application is available.

There are a lot of reasons why one may need such alternative types of
configurations - some special default params, special BeforeRunTasks
added, appropriately modified JavaParameters and so on.

Since ApplicationConfigurationType.getInstance and the type of a config
were hardcoded, so far it was needed to copy the existing logic with
appropriate modifications. It's not very convenient taking into account
that this logic may change in meantime (like lately) and one needs to
check whether the logic in a plugin also should be updated.