SBT import wizard: remove "Use auto import" checkbox
That checkbox complicates the import process for novices (and SBT project is now the default format), that settings should probably be placed in the usual global / project settings, not at the start of each project import / creation. Besides, there are known problems with auto-import (and now we have the convenient "Refresh Project" popup).