Added hydra global settings
I've added application-level settings for Hydra. This settings are
shared between all Scala projects that use Hydra as a compiler.

Also added checking if the artifacts are present in the ivy
cache. If they do not exist in the ivy cache the user is able to
download them again. In the other version this wasn't possible.