Support Scala compiler project
We need to be more robust when retrieving the Scala version from modules based on the Scala SDK. For the Scala project, the SDK is not a regular SDK, and the naming convention is not followed (e.g. SDK-2.12.2), instead it is simply called ‘starr’ (stable reference Scala release). We now go into the classpath components and retrieve the proper version.