Hydra Scala version detection heuristic fallbacks to module SDK version. Fix #37
Our heuristic for extracting the scala version from the scala-compiler jar name
may not work if the scala-compiler jar in the classpath has no version happended
(this for instance can happen when importing an sbt project with the `scalaHome`
setting set to point to a locally installed Scala distribution).

The implemented solution consists in falling back to the module SDK Scala
version if the version cannot be extracted from the scala-compiler jar name.