moved ScalaMoveClassesOrPackagesHandler to beginning of ext load list
ScalaMoveClassesOrPackagesHandler was previously moved all the way to
the end of the extension load list for refactoring.moveHandler to
be used only when ScalaMoveMembersHandler is not applicable.
This prevented it from ever being used, as other moveHandlers
(namely KotlinAwareJavaMoveClassesOrPackagesHandler) were used.

Now it is directly behind ScalaMoveMembersHandler.

(SCL-14704, SCL-12381, SCL-12364, SCL-14815) fixed