ScPackageImpl#findPackageObject should not cache null values #SCL-14266
We commonly use a wildcard import from the io.udash package object, which inherits type aliases for most commonly used classes. These very often fail to be recognized correctly, rendering `import io.udash._` unused, usually after reindexing/recompiling/updating sbt project etc. The workaround we used was to remove a definition of a method/class or an import, it worked perfectly for a while after that.

I reported this bug as SCL-14266 and this change finally seems to fix it.

#SCL-14266 fixed