Annotator tests: test specific implementation
It's not clear whether we should test isolated of full implementation. In the first case, there's a coupling between the test and the implementation, and is not clear why exactly the implementation should not be triggered sometimes. In the second case, the test is implementation-agnostic, but there might be "third-party" errors (which, though, explain why the primary implementation should be skipped).

SCL-14928