Handle MaxHeap value for sbt more carefully
Previously, the mandatory project settings for the maximum heap size
of java for sbt were appended to the vm command line regardless of what
.jvmopts/.sbtopts stated.
Now the project settings can be empty.
To still enforce a minimal maximum heap size, a hidden default value
(which has the same value as the previous default settings)
is added to the vm parameters if the maximum is not explicitly defined
or an explicitly specified minimal heap size is greater than the
hidden default.

 #SCL-15087 fixed