explicitly define return types for defaultExprProviders SCL-15577 fixed
The reason being: typingRules and expansionRules in ScalaMacroEvaluator
infer the type (and hence ClassTag) to provide to loadRules from the
types of default{Type,Expr}Providers, which don't have an explicit type
annotation. Since there is only one entry in defaultExprProviders,
scalac cleverly infers its type to be Seq[ShapelessProductArgs.type]
and off we go looking for all the possible extensions matching that
type (not many are known to be found in the wild).

kudos to Ivan Motyashov