Prepared for removal of ScalaPsiElementFactoryImpl.createType.
In the near future the ScalaPsiElementFactoryImpl.createType will be removed (it is already deprecated).
I made everything override, so we will be informed by the compiler when this is the case and we can safely remove the function.

(cherry picked from commit da9eb44dd637cb7071b0e75b00fd3bf38933c577)