Fix compatible runners search in isByteCodeIndexingActive
While opening a project, the ScalaCompilerReferenceService
decides whether to enable bytecode indexing. It is done via
CompilerIndicesSettings::isBytecodeIndexingActive

One of the requirements is that there must exist a ProjectTaskRunner
that can run project's modules and is either an SBT runner or a JPS
runner.

Unfortunately, the `hasCompatibleRunner` indices function loooks only
for the first runner that can run all the modules and then checks if
it is ans SBT or JPS runner. If the first runner can run the modules,
or the module list is empty, the lookup will stop. Then, if it's
neither a JPS or an SBT runner, the result of `hasCompatibleRunners`
will be false, even if there is a good runner further on the list.

With this change, we search through the whole list for a good runner.

YouTrack: #SCL-16768 fixed