BSP Server status Icon in Status Bar
This commit adds a BSP server control icon to Status Bar. It allows
user to kill the connection (and start a new one when a new build task
is spawned)

`toggleList` method in BspServerWidgetProvider is almost the same as
in CompileServerManager, but I decided not to extract them to a single
method, as probably there will never be a single reason to change them both.