SCL-CR-1868: TypeMismatchHighlightingTest: restore logical grouping and sequencing
Using separate classes to adjust Scala version breaks the cohesive logical grouping and sequencing of the tests.

The groups (indluding headers) and the sequence are important, given the complexity and of the tests.

Ideally, we should use annotations (both on classes and on methods), not inheritance to specify and / or adjust Scala versions.

(cherry picked from commit f4d48acc906c12450cbed2e2fb5358b9e769dabc)