Ask BSP for test execution environment
When the project is imported from SBT, it is expected that the tests
executed with IntelliJ runners will behave exactly the same as if they
would be run by BSP Server. In order to achieve this, the runner must
download some critical environment parameters from BSP server via
jvmTestEnvironment, and use them to spawn child testing process. These
parameters are:

  - classpath
  - environment variables
  - working directory
  - JVM options

The place where the parameters are set is
`BspJvmEnvironmentProgramPatcher`. Unfortunately, the
JavaProgramPatchers are executed in the Event Dispatch Thread, so it
is forbidden to do a (always blocking) BSP call inside. In order to
overcome it, the BSP call is done in BspFetchTestEnvironmentTask,
which is executed on a pool thread before each test execution. The
environment is then saved via `putUserData`, so the patcher can read
it without blocking.

Another tricky thing here is mapping from IntelliJ module to BSP
target. On the other hand, the targetId is required to receive JVM
environment via BSP. Unfortunately, there may be more than one target that matches
the module, so in cases when it's impossible to unambiguously
determine which target is the right one, a popup is shown, so it's
possible to choose from a list of potential targets. The choice is
saved as a `BspFetchTestEnvironmentTask`'s setting.