Don't try to install BspFetchTestEnvironmentTask at startup
The Scala plugin tries to install BspFetchTestEnvironmentTask on
`runConfigurationAdded` event. Unfortunately, it requires
RunConfigurationEx service to be installed, and the event happens
during the initialization of it.

Apart from that, the `beforeRunTasks.add` was previously called
uncoditionally. This caused multiple additions of the task.

(cherry picked from commit a7aa3d54aacd183290267373c96dbbb06f37f967)