Workaround for lsp4j cancellation problem
lsp4j's cancellation is implemented in that way, the
`CompletableFuture` returned from the client library contains
overriden `cancel` function:
https://github.com/eclipse/lsp4j/blob/ad99a3d77139351dbc00c2dacd388fe749652309/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/RemoteEndpoint.java#L149-L152

Unfortunately, if a CompletableFuture is mapped with `applyThen`, the
information about overriding is lost, so cancel must be called on the
original future.

In BSP-related code we modify the future a few times, so as long as
it's not fixed in intellij, we have to store the original future and
call cancel on it, instead of on derived one.