Correctly handle ill-formed BSP target's URI
in BspSelectTargetDialog prompt we present allow user to choose BSP
target's URI that will be used by the test runner. We don't present
full URI to the users, but only a shortened, simplified visible name,
that is just the `id` field's value of the URI's query (so for
`file:///tmp/project&id=abc` we present just `abc`.

Previously, in case if any of the target URI's was ill-formed,we would
skip it, because we've been calling `flatten` on a
List[Option[_]]. This could lead to indices-mismatch between targetIds
and visibleNames, and invalid result from `selectedItem` method.

Now, we don't skip anything, if the `id` field cannot be extracted, we
just present the whole URI to the user.