Generic cancellation for BSP jobs
Some BSP endpoints return `CompletableFuture`s that represend BSP jobs that are
currently running on BSP server. In order to cancel these, jobs, the `cancel`
method of this future should be called. Unfortunately, the original futures that come
from lsp4j does not support transformations well - after calling `thenApply`, the `cancel`
method of the new future does not stop the job.

Since this commit, the BspServer is converted to a proxy class that
return fixed versions of CompletableFuture