Gather logs from BSP refresh and include in the exception
Improves user experience, previously the logs were hard to find in the
UI. The exception that was thrown previously was presented in "Build
-> Sync" window did not contain any details about the failure
reason. Although the focus was set to it, so the reason of the
failure was unclear. The user had to look for the logs in the root
job, which was unintuitive. No the logs are copied to the
Bsp4JJobFailure exception and presended in the focused view.