Subscribe to RunManagerListener topic manually
We have observed that sometimes the "Use BSP environment" is not added
to run configurations. Interestingly, the problem occurs during whole
IntelliJ process lifetime. Restart is usually a valid workaround.
The reproducibility ratio is very low, but one of the hypothesis is
that, the listener definied via XML file is not correctly subscribed
to the topic. That's why I'd like to directly subscribe it via
`subscribe` method call instead of defining it in XML file.