Translate CancellableFuture future to java.
In this class, we utilize the newIncompleteFuture method which was
introduced in Java 9. Neverthe less, we'd like the code to be
compilable with all JVM versions, so it's better to write it in java,
because java's override keyword is not obligatory when overriding a
method, contrary to scala.

(cherry picked from commit 9a55cb5bb5565efab304c7d4e83d6038df219159)