Import BSP project via Fastpass
This commit implements import of Pants projects to IntelliJ by converting
them to BSP projects with fastpass.

The tricky part was that, contrary to SBT-BSP imports, fastpass creates
BSP configuration *outside* of project root. This means, if the repository
root is `~/pants-example` and one wants to import `~/pants-example/java_app`,
the BSP configuration is not created in `~/pants-example/.bsp`, but in
`~/bsp-projects/java_app/.bsp`.

In order to support this case, the field `externalBspWorkspace` is introduced,
where BSP configuration is generated and stored, even if something else was chosen
by the user to import. Despite that, the BspSetupConfigStep must still receive
the original path selected by the user, because it needs to know what exactly is
supposed to be exported.

Thanks to this behavior, a user can import a few small, independent subprojects simultaneously
without touching rest of the project.

References:
https://github.com/scalameta/fastpass