TastyReader: don't re-read .tasty files from JARs during indexing, #SCL-18428
Extracting every .tasty file from a JAR separately can result in non-linear time complexity. Given that, during indexing, we already have the required data in a byte array, we can save that data to a file directly, which should improve batch indexing.

This is, however, not an issue for the Go To and Type Info actions.