TastyReader: check binary signature and version of .tasty files during indexing, #SCL-18428
This prevents instantiating a TASTy reader instance unnecessarily, for non-TASTy files with .tasty extensions and for TASTy files that have incompatible version (and which we cannot read anyway).

This is potentially a significant optimization, because TASTy reader is (effectively) an instance of the Scala 3 compiler, and the required JARs might not yet be downloaded.