TastyReader: bundle libraries instead of downloading them (potentially during indexing), #SCL-18428
Downloading TASTy inspector libraries is a trade off. However, during a Go To or Type Info action, those libraries are likely to be already downloaded, which is often not the case for indexing.

The "official" TASTy reader depends on parts of the Scala 3 compiler, which is around 14 MB. On the one hand, that's a lot. On the other hand, the Scala plugin is already 84 MB (uncompressed), and already includes the Scala 2 compiler library.

Because indexing is a rather important functionality, let's pre-bundle the required JARs for now (that would be ~10% of the plugin size). In any case, we need to find a more lightweight method to process TASTy during indexing (and not just because of the JARs size).

What's more, we need to find some way to minimize the ZIP size - ~100 MB is a lot for a plugin, regardless of the TASTy reader JARs (particularly - why do we bundle the Scala 2 compiler?).