Enable github workflow on idea211.x branch
This is an empty commit that mimics
38914a0afcfea8db1ad4637df6cddc723a216c59,
just so that our branch-checking/cherry-picking
knows that the commit has been cherry-picked.
The original cherry-pick that was supposed to do this
(88124b027d6380ff40435f7c8e1cb8e6ff892292)
failed in that, because a bug in cherry-up
screwed up the commit message (fixed in cherry-up v1.1).
The actual content of the commit was, however, successfully
cherry-picked by the first attempt, hence why this commit is empty.

(cherry picked from commit 38914a0afcfea8db1ad4637df6cddc723a216c59) -
empty cherry pick