Implement a warning for when a private field or a class parameter shadows a superclass field #SCL-20050 Fixed
This commit contains a number of related changes:
1. `FieldShadowingInspection` - a new inspection checking if a local field's name shadows a name of a class
or trait field name of the encompassing class or parent classes, but in the case of parent classes we need
to check the visibility modifier - private fields of parent classes should be ignored as they are not visible
from the place of the field under inspection.
2. There is a corner case where a class field can shadow a variable from the parent class. If the parent class
field was an immutable value, it would be overriden (so, not shadowed) but if it's a mutable variable, it's not
overriden but shadowed instead. It's a limitation of the Scala compiler. By default, the compiler does not warn
about this corner case. The warning can be turned on with the compiler's option `-Xlint:private-shadow`.
The inspection also handles this corner case.
3. The way the corner case about shadowing a variable from a parent class is handled is controlled by a new widget
in the inspection configuration view - a combobox with three options: "Always highlight", "Never highlight", and
"Check the compiler option". The default option is "Always". The third option means that the inspection will look
to the list of the Scala compiler options and will highlight the corner case or not depending on if `-Xlint:private-shadow`
is used or not. This is the first time a combobox is used in this way, so it's a bit of additional code.
4. There are two quickfixes for the inspection: Rename the shadowing field, and disable the inspection
(it is possible that some users do it on purpose).