remove ScalaDocumentationFromJavaTest #SCL-20088 fixed
Originally the test was written with message
"scaladoc: added test for calling scaladoc from java code #SCL-8760"
But the test requires accessing platform API which is not meant to be public. It's deprecated and will be removed later.
We shouldn't test this platform logic.
Platform is responsible for choosing the correct DocumentationProvider for the given element