Implement a combobox option for highlighting only the -Xlint:private-shadow corner case
I tested this inspection with different projects and it looks like some of the teams don't care about shadowing
class/trait fields by local fields but may be still interested in warning against the corner case controlled by
the compiler option -Xlint:private-shadow. So I implemented a new option for the combobox. If it's chosen,
the inspection will highlight **only** the suspicious shadowing of a mutable variable by a subclass field, but
not the more typical cases of shadowing a class/trait field by a local field.

I made changes to some field names, to simplify them.
And I removed the file `FieldShadowing.html` which was replaced by `FieldShadow.html` but waa not removed before.