ScalaTestAstTransformer.transformNode: do not check for ref spec redundantly when constructing AST Tree for scalatest-finders, fix FlatSpecFinder
there was an issue with FlatSpecFinder:
When you click on test class name, it creates run configuration for the "whole suite", which contains just the test class name. But if you click inside test class (e.g. on opening brace) it creates run configuration which lists all test names.
It's not good:
1. It is inconsistent between all other finders, which just return null selection for such cases (though we might have considered returning selection which would select only class, not to fallback to ScalaTestSingleTestLocationFinderOld but it's a different story)
2. If the test class contain tests with dynamic test name, they won't be detected

the change was done within review for #SCL-20143 (SCL-CR-265)