Reusable Combobox for inspection options #SCL-20201
This is a generalization of the logic used already in FieldShadowInspection.
A combobox in the inspection's configuration view controls a flag that the inspection uses in its logic.
The combobox lets the user switch between "enabled", "disabled", and "check the compiler option".
In the last case the compiler option's name needs to be provided and then the flag is set to true if the compiler option is used.
It is also possible to compute the flag's value in a more complex way where the logic can check the element under inspection.

I updated our checkbox implementation in a similar way and then used both in FieldShadowInspection, ScalaUnusedDeclarationInspection and NonLocalReturnInspection.