Remove quickfixes from NonLocalReturnInspection #SCL-20250 Fixed
The two quickfixes already in use can break code semantics = they are removed.
Also, the default mode of the inspection will be now to run only if the compiler option `-Xlint:nonlocal-return` is present.
The user may switch it in the configuration view.

Please note that the way the checkbox is implemented will be changed shortly, but not in this PR.