Adding a checkbox to comply with -Xfatal-warnings and/or -Werror
In Scala 2.13 we have a compiler option -Werror which makes the compiler treat warning as error. In Scala 3 its name is changed to -Xfatal-warnings
I added a checkbox to the configuration panel - if set, we will check the compiler options for any of these two and if they are present,
we will  display "private shadow" as an error.

Both checkboxes about complying with compiler options are checked by default.

Because the scope of the inspection changed, I also made some changes in variable names and the inspection's labels.