SCL-19526 Fix SOE in implicit search with raw types + f-bounds
Because such types are represented with ScExistentialArgument.Deferred args
(which don't / can't have a useful equals), the recursion guard was being
bypassed.

Replace such Deferred args with dummy stand-in types comprising the
corresponding original class type parameter and the bounds, and use this
in the recursion guard instead.

I've also added a more efficient path for unbounded type parameters;
these can't cause cycles and so can use Complete existential args.
This is not functionally needed by the test. By itself, it would
fix this simpler version of the test:

```diff
-interface Column<V extends String>
+interface Column<V>
```