copy/paste: don't try guessing which imports to add when copying java code from raw text file to scala file
Remove entire AdditionalImportsResolver.scala - `handleReference` method lead to assertion exception:
```
java.lang.IllegalStateException: Must be created in a background thread under the read action
	at com.intellij.codeInsight.daemon.impl.quickfix.ImportClassFixBase.<init>(ImportClassFixBase.java:58)
```
this assertion was added in 223.x.
So we shouldn't do any resolve in during pasting, which is done on EDT.
Maybe we could somehow resolve references by showing progress dialog and resolving in the background...
But I thought that it will be fine not to try guessing.
Even Java doesn't do that when copy/paste Text (with Java code) to actual Java code. It doesn't add any guessed imports.