drop "testDiscovery" ("Show Affected Tests" action) #SCL-14937
It's in some old, unmaintained state. "TestDiscovery" support was in some intermediate state and wasn't finished. Nobody actually used it. Currently the code just uses some deprecated API and complicates project structure.
If we decide to implement "Show Affected Tests" in the future we should approach to it from the scratch