Type parameter shadowing doesn't work in Scala 3 #SCL-20255 Fixed
Checking for suspicious type parameter shadowing was disabled in code by SCL-19723 because the highlight showed also in case of type lambdas.
Now the bug in type lampdas is fixed and the inspection no longer treats it as suspicious shadowing.
It means that it was enough to enable the inspection back for Scala 3 to fix SCL-20255.